+++ aws_account/02_core/main.tf
@@ -125,19 +125,19 @@ module "alb_public" {
   depends_on         = [module.route53]
 }
 
-# Internal Load Balancer (Private)
-module "alb_private" {
-  source             = "../../modules/load_balancer"
-  name               = "alb-private" # Keep this name
-  public             = false
-  vpc                = module.myvpc.vpc
-  subnets            = module.myvpc.subnets_private
-  root_domain        = module.route53.domain
-  log_bucket         = data.aws_s3_bucket.logging
-  log_retention_days = var.log_retention_days
-  tags               = var.tags
-  depends_on         = [module.route53]
-}
+# # Internal Load Balancer (Private)
+# module "alb_private" {
+#   source             = "../../modules/load_balancer"
+#   name               = "alb-private" # Keep this name
+#   public             = false
+#   vpc                = module.myvpc.vpc
+#   subnets            = module.myvpc.subnets_private
+#   root_domain        = module.route53.domain
+#   log_bucket         = data.aws_s3_bucket.logging
+#   log_retention_days = var.log_retention_days
+#   tags               = var.tags
+#   depends_on         = [module.route53]
+# }
 
 # Setting up a VPN has a fairly high cost
 # You probably don't need it and if you do, it can be enabled temporarily