+++ aws_account/03_apps/bastion.tf
@@ -52,32 +52,62 @@ data "aws_iam_instance_profile" "ec2" {
 #   tags                 = var.tags
 # }
 
-# Bastion - Debian 12 Bookworm x86_64
-module "bastion_debian12_x86_64" {
-  source               = "../../modules/ec2"
-  name                 = "bastion"
-  subnet_id            = data.aws_subnets.private.ids[0]
-  os                   = "debian12"
-  arch                 = "x86_64"
-  machine              = "t3a.small"
-  security_groups      = [data.aws_security_group.main.id]
-  iam_instance_profile = data.aws_iam_instance_profile.ec2
-  userdata             = "../../userdata/debian.sh"
-  kms_key              = data.aws_kms_key.main
-  tags                 = var.tags
-}
+# # Bastion - Debian 12 Bookworm x86_64
+# module "bastion_debian12_x86_64" {
+#   source               = "../../modules/ec2"
+#   name                 = "bastion"
+#   subnet_id            = data.aws_subnets.private.ids[0]
+#   os                   = "debian12"
+#   arch                 = "x86_64"
+#   machine              = "t3a.small"
+#   security_groups      = [data.aws_security_group.main.id]
+#   iam_instance_profile = data.aws_iam_instance_profile.ec2
+#   userdata             = "../../userdata/debian.sh"
+#   kms_key              = data.aws_kms_key.main
+#   tags                 = var.tags
+# }
 
-# Bastion - Debian 11 Bullseye x86_64
-module "bastion_debian11_x86_64" {
-  source               = "../../modules/ec2"
-  name                 = "bastion"
-  subnet_id            = data.aws_subnets.private.ids[0]
-  os                   = "debian11"
-  arch                 = "x86_64"
-  machine              = "t3a.small"
-  security_groups      = [data.aws_security_group.main.id]
-  iam_instance_profile = data.aws_iam_instance_profile.ec2
-  userdata             = "../../userdata/debian.sh"
-  kms_key              = data.aws_kms_key.main
-  tags                 = var.tags
-}
+# # Bastion - Debian 11 Bullseye x86_64
+# module "bastion_debian11_x86_64" {
+#   source               = "../../modules/ec2"
+#   name                 = "bastion"
+#   subnet_id            = data.aws_subnets.private.ids[0]
+#   os                   = "debian11"
+#   arch                 = "x86_64"
+#   machine              = "t3a.small"
+#   security_groups      = [data.aws_security_group.main.id]
+#   iam_instance_profile = data.aws_iam_instance_profile.ec2
+#   userdata             = "../../userdata/debian.sh"
+#   kms_key              = data.aws_kms_key.main
+#   tags                 = var.tags
+# }
+
+# # Bastion - Rocky Linux x86_64
+# module "bastion_rocky9_x86_64" {
+#   source               = "../../modules/ec2"
+#   name                 = "bastion"
+#   subnet_id            = data.aws_subnets.private.ids[0]
+#   os                   = "rocky9"
+#   arch                 = "x86_64"
+#   machine              = "t3a.small"
+#   security_groups      = [data.aws_security_group.main.id]
+#   iam_instance_profile = data.aws_iam_instance_profile.ec2
+#   userdata             = "../../userdata/rhel.sh"
+#   kms_key              = data.aws_kms_key.main
+#   tags                 = var.tags
+# }
+
+# # Bastion - Rocky Linux ARM64
+# module "bastion_rocky9_arm64" {
+#   source               = "../../modules/ec2"
+#   name                 = "bastion"
+#   subnet_id            = data.aws_subnets.private.ids[0]
+#   os                   = "rocky9"
+#   arch                 = "arm64"
+#   machine              = "t4g.small"
+#   security_groups      = [data.aws_security_group.main.id]
+#   iam_instance_profile = data.aws_iam_instance_profile.ec2
+#   userdata             = "../../userdata/rhel.sh"
+#   kms_key              = data.aws_kms_key.main
+#   tags                 = var.tags
+# }